requirements.txt
matplotlib==3.3.4
pillow==9.1.0
tensorflow==2.4.2
opencv-python==4.2.0.34
numpy==1.19.5
pandas==1.1.5
seaborn==0.11.2
keras==2.4.2
tensorflow_addons==0.12.0
openpyxl==3.0.9
scikit-learn==1.0.2
tqdm==4.62.3
geocoder==1.38.1
